+++ pkg/local_object_storage/metabase/control.go
@@ -259,11 +259,6 @@ func (db *DB) ResyncFromBlobstor(bs common.Storage, shardID string, onIterationE
 		return fmt.Errorf("could not flush remaining objects to metabase: %w", err)
 	}
 
-	err = db.SyncCounters()
-	if err != nil {
-		return fmt.Errorf("could not sync object counters: %w", err)
-	}
-
 	return nil
 }